+++ .env-example
@@ -1,5 +1,5 @@
 NODE_ENV=dev
-LOG_LEVEL=debug
+LOG_LEVEL=error
 LOG_FILE=/var/log/clive.log
 DB_FILE=/srv/clive/db.json
 DISCORD_WEBHOOK_URL=https://discordapp.com/api/webhooks/YOUR_DISCORD/WEBHOOK_URL_HERE
