+++ .eslintrc
@@ -7,7 +7,8 @@
       "error",
       {
         "singleQuote": true,
-        "trailingComma": "all"
+        "trailingComma": "all",
+        "endOfLine": "auto"
       }
     ]
   },
