+++ package.json
@@ -5,27 +5,27 @@
 	"license": "ISC",
 	"main": "index.js",
 	"scripts": {
-		"start": "nodemon index.js",
-		"dev": "nodemon --inspect index.js",
+		"start": "nodemon src/index.js",
+		"dev": "nodemon --inspect src/index.js",
 		"precommit": "lint-staged"
 	},
 	"dependencies": {
 		"dotenv": "^8.2.0",
 		"lodash": "^4.17.15",
 		"lowdb": "^1.0.0",
-		"nodemon": "^2.0.2",
+		"nodemon": "^2.0.4",
 		"request": "^2.88.2",
 		"request-promise": "^4.2.5",
-		"twitch-js": "^2.0.0-beta.30",
+		"twitch-js": "^2.0.0-beta.33",
 		"winston": "^3.2.1"
 	},
 	"devDependencies": {
-		"eslint": "^6.8.0",
-		"eslint-config-prettier": "^6.10.0",
-		"eslint-plugin-prettier": "^3.1.2",
-		"husky": "^4.2.3",
-		"lint-staged": "^10.0.8",
-		"prettier": "^1.19.1",
+		"eslint": "^7.0.0",
+		"eslint-config-prettier": "^6.11.0",
+		"eslint-plugin-prettier": "^3.1.3",
+		"husky": "^4.2.5",
+		"lint-staged": "^10.2.2",
+		"prettier": "^2.0.5",
 		"prettier-eslint": "^9.0.1"
 	},
 	"lint-staged": {
